requirements.txt
Django==2.2.1
django-cors-headers==3.0.1
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
pytz==2019.1